Fix "[n lines truncated]" lines being coloured
They're supposed to be grey.

The format pattern for the n-lines-truncated messages was missing an
identifier for the Bash colour reset token. Moreover, the formattedWrite
call for when lines had been printed (vh -nx for x > 0) was printing the
filename when it should have been printing the file/files plurality.
The call for when lines had not been printed (vh -n0) was outright
malformed, and this could have been caught at compile-time had we used
formattedWrite!pattern(...). That would mean not supporting GDC.